restructuring tests, fix bootstrap test, started storage test
