+++ tests/Integration/StreamContext/Stream.php
@@ -1,6 +1,4 @@
 <?php
-use PHPUnit\Framework\TestCase;
-
 class Stream
 {
     protected $baseuri;
